services/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
mysql-connector-python==8.3.0
python-telegram-bot==21.0.1
requests==2.31.0